% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccsbt_effort_level0.R
\name{ccsbt_effort_level0}
\alias{ccsbt_effort_level0}
\title{Extract source georeferenced efforts datasets of CCSBT from the Tuna atlas database}
\usage{
ccsbt_effort_level0(year_tunaatlas)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}
}
\description{
This function extracts the source georeferenced effort datasets stored in the Tuna atlas database coming from the Commission for the Conservation of Southern Bluefin Tuna (CCSBT). The output dataset provides efforts targetting tuna, tuna-like and shark species in the area managed under the jurisdiction of the CCSBT. Efforts are stratified by year, month, fishing gear, fishing country, unit of effort, area (5° squares). Data are expressed using CCSBT's coding system.
}
\details{

}
\examples{

# Retrieve ccsbt georeferenced efforts data from 2017
ccsbt_effort<-ccsbt_effort_level0(2017)
head(ccsbt_effort)

}
\seealso{
Other process data: \code{\link{ccsbt_datasets_level0}},
  \code{\link{convert_units}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{iattc_datasets_level0}},
  \code{\link{iccat_catch_level0}},
  \code{\link{iccat_effort_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{iotc_effort_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}},
  \code{\link{wcpfc_catch_level0}},
  \code{\link{wcpfc_effort_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
