% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar.R
\name{create_calendar}
\alias{create_calendar}
\title{Creates a continious calendar}
\usage{
create_calendar(firstdate,finaldate,resolution,unit)
}
\arguments{
\item{firstdate}{character string first date of the calendar, format : YYYY-MM-DD}

\item{finaldate}{character string final date of the calendar, format : YYYY-MM-DD}

\item{resolution}{real temporal resolution of the calendar in day, month or year (see: temporal_reso_unit). Note: for 1/2 month put temporal_reso=1/2 and temporal_reso_unit="month" , type = integer;}

\item{unit}{character string temporal resolution unit. accepted value : "day" "month" "year"}
}
\value{
A dataframe with two columns : "time_start" (starting date of the period) and "time_end" (ending date of the period).
}
\description{
Creates a continious calendar according to input parameters : first and final date of the calendar, temporal resolution and temporal resolution unit
}
\details{
Values for unit can be: "day","month","year". 

The algorithm can create a mi-month calendar : first period between 1st and 15th of a month, second period between 16th and the end of the month. For this set resolution=0.5 and unit="month"
}
\examples{

calendar<-create_calendar("1900-01-01","2020-01-01",15,"day")
calendar<-create_calendar("1900-01-01","2020-01-01",0.5,"month")

}
\seealso{
Other process data: \code{\link{convert_units}},
  \code{\link{create_grid}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{rasterize_geo_timeseries}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}}
}
\author{
Chloé Dalleau, \email{dalleau.chloe@hotmail.fr}; modified by Paul Taconet, \email{paul.taconet@ird.fr}
}
