% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Creates a continious grid}
\usage{
create_grid(latmin,latmax,lonmin,lonmax,resolution,crs,centred)
}
\arguments{
\item{latmin}{real smallest latitude for the spatial grid in degree (range: -90 to 90)}

\item{latmax}{real biggest latitude wanted for the spatial grid in degree (range: -90 to 90)}

\item{lonmin}{real smallest longitude for the spatial grid in degree (range: -180 to 180)}

\item{lonmax}{real biggest longitude wanted for the spatial grid in degree (range: -180 to 180)}

\item{resolution}{real spatial resolution of the grid in degree}

\item{crs}{character string a character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation.}

\item{centred}{boolean TRUE for a grid centred on zero. default is TRUE}
}
\value{
A sf object (package : sf) based on input parameters and composed by square polygons.
}
\description{
Create a continious (regular) geospatial grid
}
\details{
The default CRS is EPSG 4326: "+init=epsg:4326 +proj=longlat +datum=WGS84"
}
\examples{

grid <- create_grid(latmin=-90,latmax=90,lonmin=-180,lonmax=80,resolution=1,centred=T)

}
\seealso{
Other process data: \code{\link{convert_units}},
  \code{\link{create_calendar}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{rasterize_geo_timeseries}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}}
}
\author{
Chloé Dalleau, \email{dalleau.chloe@hotmail.fr}; modified by Paul Taconet, \email{paul.taconet@ird.fr}
}
