% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection_sardara_france.R
\name{db_connection_sardara_france}
\alias{db_connection_sardara_france}
\title{Connects to Sardara France PostgreSQL database stored on the Observatory of Exploited Tropical Pelagic Ecosystems (Ob7) servers.}
\usage{
db_connection_sardara_france(db)
}
\value{
a wrapper of RPostgreSQL connection (object of type "con")
}
\details{
A VPN connection is necessary to connect to this database.
}
