% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection_sardara_world.R
\name{db_connection_sardara_world}
\alias{db_connection_sardara_world}
\title{Connects to Sardara World PostgreSQL database stored on the BlueBridge H2020 project servers.}
\usage{
db_connection_sardara_world(db)
}
\value{
a wrapper of RPostgreSQL connection (object of type "con")
}
