% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connexion.R
\name{db_connexion}
\alias{db_connexion}
\title{Connects to a postgreSQL database and returns an object of type "con".}
\usage{
db_connexion(db)
}
\description{
The databases connexions parameters must be set in this document: \url{https://raw.githubusercontent.com/ptaconet/rtunaatlas/master/inst/extdata/db_connexions_parameters.csv}
The column "connexion_name_package" is the name of the connexion used in the package rtunaatlas
}
