% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_load_dataset_in_db.R
\name{functions_load_dataset_in_db}
\alias{functions_load_dataset_in_db}
\alias{FUNMergeDimensions_CodeListLike}
\title{Functions used to load a dataset in a database with Sardara model}
\usage{
FUNMergeDimensions_CodeListLike(DBconnection, DBTableName, DB_PK_AttributeName,
  DB_CodeRFMO_AttributeName, DB_TableRFMO_AttributeName, inputDataset,
  HarmonizedDF_CodeRFMO_AttributeName, DB_Tables_ToLookInto)
}
\description{
A set of functions used to load a dataset in the database
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
