% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_visualize_data.R
\name{functions_visualize_data}
\alias{functions_visualize_data}
\alias{fun_aggregate_keep_n_classes}
\title{Functions used to load a dataset in a database with Sardara model}
\usage{
fun_aggregate_keep_n_classes(df_plot, number_of_classes, ColnameVariable)
}
\description{
A set of functions used to vizualise datasets
}
\seealso{
Other visualize data: \code{\link{pie_map}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
