% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_metadata.R
\name{generate_metadata}
\alias{generate_metadata}
\title{Generate the dataset of metadata}
\usage{
generate_metadata(metadata_file,dataset_type)
}
\arguments{
\item{metadata_file}{data.frame of metadata (input)}

\item{dataset_type}{string type of dataset (raw_dataset, codelist, mapping)}
}
\value{
a data.frame of metadata with the structure to use as input of the functions to load a dataset
}
\description{
Generate a dataset of metadata used as input of the functions to load a dataset in the database from a template of metadata dataframe
}
\details{

}
\examples{


}
\seealso{
Other load data: \code{\link{functions_load_dataset_in_db}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
