% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dsd_of_dataset.R
\name{get_dsd_of_dataset}
\alias{get_dsd_of_dataset}
\title{Get the data structure definition of a raw_dataset}
\usage{
get_dsd_of_dataset(con, dataset_metadata)
}
\description{
This function creates the data structure definition of a raw_dataset stored in the database.
}
\seealso{
Other discover data: \code{\link{get_codelist_of_dimension}},
  \code{\link{list_dataset_available_dimensions}},
  \code{\link{list_metadata_datasets}},
  \code{\link{list_variable_available_dimensions}}
}
