% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSQLSardaraQueries.R
\name{getSQLSardaraQueries}
\alias{getSQLSardaraQueries}
\title{Get a list of queries to execute to plug services to the datasets available in Sardara}
\usage{
getSQLSardaraQueries(con,dataset_metadata)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database)}

\item{dataset_metadata}{data.frame of type "metadata" (one row extracted from the table metadata.metadata).}
}
\value{
a list with SQL queries and informations on the dataset. Depending on the type of data (raw_dataset, codelist, mapping), the output elements of the list might vary. The elements of the list are:
\itemize{
 \item{\code{query_CSV} :}{ Query to retrieve a data.frame with the codes}
 \item{\code{query_CSV_with_labels} :}{ Query to retrieve a data.frame with codes and labels (for raw_datasets)} 
 \item{\code{query_NetCDF} :}{ }
 \item{\code{query_wfs_wms} :}{ }
 \item{\code{query_wfs_wms_aggregated_layer} :}{ }
 \item{\code{query_dynamic_list_keywords_species} :}{ }
 \item{\code{query_dynamic_list_keywords_fishing_gears} :}{ }
 \item{\code{query_dynamic_list_keywords_Fleets} :}{ }
 \item{\code{query_dynamic_metadata_count_features} :}{ }
 \item{\code{query_dynamic_metadata_spatial_Extent} :}{ }
 \item{\code{query_dynamic_metadata_temporal_Extent} :}{ }
 \item{\code{query_dynamic_metadata_get_SRID} :}{ }
 \item{\code{lineage} :}{ }
 \item{\code{query_dynamic_list_keywords_institutions} :}{ }
  }
}
\description{
Get a list of queries to execute to manage the datasets available in Sardara (extract in csv format, in netcdf, etc.)
}
\examples{

 con=db_connection_sardara_world()
 
 # retrieve metadata row of dataset global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1
dataset_metadata<-list_metadata_datasets(con,dataset_name="global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1")
queries<-getSQLSardaraQueries(con,dataset_metadata)

# retrieve data.frame of global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1

global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1<-dbGetQuery(con,queries$query_CSV)

}
\seealso{
Other access data: \code{\link{extract_dataset}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
