% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_time_structure.R
\name{harmonize_time_structure}
\alias{harmonize_time_structure}
\alias{format_time_db_format}
\title{Set of functions to harmonize the time dimension structure of the source tRFMOs datasets}
\usage{
harmo_time_1(df_input,colname_year,colname_timeperiod)
harmo_time_2(df_input,colname_year,colname_month)
harmo_time_3(df_input,colname_year,colname_monthstart,colname_monthstop)
}
\arguments{
\item{df_input}{data.frame with the time structure to harmonize}

\item{colname_year}{string. Name of the column of years.}

\item{colname_timeperiod}{string. Name of the column of period of time.}

\item{colname_month}{string. Name of the column of month.}

\item{colname_monthstart}{string. Name of the column of starting month of validity of the data.}

\item{colname_monthstop}{string. Name of the column of ending month of validity of the data}
}
\value{
a data.frame in which time dimension is structured with two columns:
\itemize{
\item{time_start}: {providing the starting date of validity of the data (included)}
\item{time_end}: {providing ending date of validity of the data (excluded)}
 }
}
\description{
Set of functions to harmonize the structure of the datasets for the time dimension
}
\details{
\itemize{
\item{colname_year}: {Integer: YYYY}
\item{colname_timeperiod}: {Integer. Periods are defined as following:}
\itemize{
\item{1 to 12}: {1 month time period. The number gives the month number in the year (e.g. 1=January, 2=February, etc.);}
\item{13 to 16}: {1 quarter time period. 13=January to March, 14=April to June, 15=July to September, 16=October to December;}
\item{17}: {One year time period;}
\item{18 and 19}: {1 semester time period. 18=January to June, 19=July to December}
}
\item{colname_month}: {Integer: MM}
\item{colname_monthstart}: {Integer: MM}
\item{colname_monthstop}: {Integer: MM}
}

In 2017, the time structure of the source tRFMOs datasets are converted with the following functions:
\itemize{
\item{harmo_time_1}: {}
\item{harmo_time_2}: {}
\item{harmo_time_3}: {}
}

The functions harmo_time_1, harmo_time_2, harmo_time_3, harmo_time_4 work as following: 

1) Fist they convert the source time structure to the following target structure:
\itemize{
\item{One column "Year" giving the year,}
\item{One coumn "MonthStart" giving the first month of the period,}
\item{One column "Period" giving the period. The period is defined as hereunder:}
 \itemize{
\item{1 = one month}
\item{3 = one trimester}
\item{6 = one semester}
\item{12 = one year}
}}

2) Then they convert the latter structure to the structure with the time_start and time_end columns (this step is done using the function \code{format_time_db_format})
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
