% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iattc_catch_level0.R
\name{iattc_catch_level0}
\alias{iattc_catch_level0}
\title{Extract source georeferenced catch datasets of IATTC from Sardara World database}
\usage{
iattc_catch_level0(year_tunaatlas)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}

\item{raise_flags_to_typeofschool}{boolean.}

\item{dimension_to_use_if_no_raising_flags_to_typeofschool}{string.}
}
\description{
This function extracts the source georeferenced catch datasets stored in the Sardara World database coming from the  (IATTC). The output dataset provides the catch of tuna, tuna-like and by-catch in the area managed under the jurisdiction of the iattc. Catches are stratified by month, species, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (usualy 1째 or 5째 square) and unit of catch (weight or number). Data are expressed using IATTC's coding system.
}
\details{
The output dataset lists catch of tuna, tuna-like and shark species in the Eastern Atlantic ocean. Catches are stratified by month, species, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (1째 or 5째 square) and unit of catch (weight or number). This dataset is computed using public domain datasets released by the Indian Ocean Tuna Commission (iattc).
This function merges the primary catch-and-effort datasets released by IATTC.
Data are expressed using IATTC's coding system.
}
\examples{

# Retrieve IATTC georeferenced catch data from 2017, with dataset with flag dimension raised to dataset with schoolytpe dimension
iattc_catch<-iattc_catch_level0(2017,raise_flags_to_typeofschool=TRUE)
head(iattc_catch)

}
\seealso{
Other extract data: \code{\link{ccsbt_catch_level0}},
  \code{\link{extract_dataset}},
  \code{\link{iccat_catch_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{iotc_effort_level0}},
  \code{\link{wcpfc_catch_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
