% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iattc_datasets_level0.R
\name{iattc_datasets_level0}
\alias{iattc_datasets_level0}
\title{Extract primary IATTC datasets from the Tuna atlas database}
\usage{
iattc_datasets_level0(variable,year_tunaatlas,raise_flags_to_schooltype,dimension_to_use_if_no_raising_flags_to_schooltype)
}
\arguments{
\item{variable}{string. Variable to extract. accepted values: "catch", "effort"}

\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}

\item{raise_flags_to_schooltype}{boolean. Raise dataset with flag stratification to dataset with schooltype stratification? See section Details for more information.}

\item{dimension_to_use_if_no_raising_flags_to_schooltype}{NULL or string. Not nullable if \code{raise_flags_to_schooltype} is set to FALSE. if not NULL, either set to "flag" or "schooltype"}
}
\description{
This function extracts the primary datasets of geo-spatial gridded catch or efforts coming from the Inter American Tropical Tuna Commission (IATTC) and stored in the Tuna atlas database.
}
\details{
The output dataset lists catch or effort of tuna, tuna-like and shark species in the Eastern Atlantic ocean. Catches or efforts are usually stratified by year, month, species (for catches only), fishing gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (1° or 5° square) and unit. Some of these dimensions can be missing depending on the confidentialy policies of each RFMO. The output dataset is computed using public domain datasets released by the IATTC.
Primary longline and pole-and-line datasets are simply merged.

Details on the use of the parameter \code{raise_flags_to_schooltype}: For confidentiality policies, information on flag and school type for the geo-referenced catches is available in separate files for Purse seine datasets.
\itemize{
\item{ If the parameter \code{raise_flags_to_schooltype} is set to TRUE, for each stratum, the catch/effort from the flag-detailed dataset will be raised to the catch/effort from the school type-detailed dataset to get an estimation of the catches by flag and school type in each stratum.}
\item{ If the parameter \code{raise_flags_to_schooltype} is set to FALSE, one single dataset will be used and in this case, the parameter \code{dimension_to_use_if_no_raising_flags_to_schooltype} must be filled in: }
 \itemize{
\item{ If the parameter \code{dimension_to_use_if_no_raising_flags_to_schooltype} is set to "flag", the data with flag information will be used.}
\item{ If the parameter \code{dimension_to_use_if_no_raising_flags_to_schooltype} is set to "schooltype", the data with schooltype information will be used. }
}
}



The output dataset is expressed with IATTC's coding system.
}
\examples{

# Retrieve IATTC georeferenced catch data from 2017, with dataset with flag dimension raised to dataset with schoolytpe dimension
iattc_catch<-iattc_catch_level0("catch",2017,raise_flags_to_schooltype=TRUE)
head(iattc_catch)

}
\seealso{
Other process data: \code{\link{ccsbt_datasets_level0}},
  \code{\link{ccsbt_effort_level0}},
  \code{\link{convert_units}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{iccat_catch_level0}},
  \code{\link{iccat_effort_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{iotc_effort_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}},
  \code{\link{wcpfc_catch_level0}},
  \code{\link{wcpfc_effort_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
