% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_iattc_functions.R
\name{iattc_functions}
\alias{iattc_functions}
\alias{create_additional_columns_IATTC_CE}
\title{Functions to harmonize the structure of the IATTC datasets}
\usage{
create_additional_columns_IATTC_CE(input_IATTC_CE_file, ColLat, ColLon,
  SquareSizeDegrees, GearCode)
}
\description{
Set of functions to harmonize the structure of the IATTC datasets
}
\seealso{
Other harmonize data structure: \code{\link{harmonize_time_structure}},
  \code{\link{iccat_functions}},
  \code{\link{iotc_functions}},
  \code{\link{wcpfc_functions}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
