% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccat_catch_level0.R
\name{iccat_catch_level0}
\alias{iccat_catch_level0}
\title{Extract source georeferenced catch datasets of ICCAT from Sardara World database}
\usage{
iccat_catch_level0(year_tunaatlas,include_type_of_school)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}

\item{include_type_of_school}{boolean. Set TRUE if you want the output dataset with school type stratification. FALSE will provide the stratification without the type of school. See details for more information.}
}
\description{
This function extracts the source georeferenced catch datasets stored in the Sardara World database coming from the International Commission for the Conservation of Atlantic Tuna (ICCAT). The output dataset provides the catch of tuna, tuna-like and by-catch in the area managed under the jurisdiction of the ICCAT Catches are stratified by month, species, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (usualy 1째 or 5째 square) and unit of catch (weight or number).
}
\details{
The output dataset lists catch of tuna, tuna-like and shark species in the Atlantic ocean. Catches are stratified by month, species, gear, vessel flag reporting country, fishing mode (i.e. type of school used) - if parameter \code{include_type_of_school} is set to TRUE, area (1째 or 5째 square) and unit of catch (weight or number). This dataset is computed using public domain datasets released by the International Commission for the Conservation of Atlantic Tuna (ICCAT).

Details on the use of the parameter \code{include_type_of_school}: ICCAT delivers two catch-and-efforts datasets for purse seiners: one that gives the detail of the type of school (Fad|Free school) for purse seine fisheries and that starts in 1994 (called Task II catch|effort by operation mode Fad|Free school) and one that does not give the information of the type of school and that covers all the time period (from 1950) (called Task II catch|effort). These data are redundant (i.e. the data from the dataset Task II catch|effort by operation mode are also available in the dataset Task II catch|effort) but in the latter, the information on the type of school is not available.
\itemize{
\item{ If the parameter \code{include_type_of_school} is set to TRUE, both datasets will be combined to produce a dataset that covers the whole time period, with fishing mode information (Fad | free school).}
\item{ If the parameter \code{include_type_of_school} is set to FALSE, only the dataset without the type of school information will be used. The output dataset will hence not have the stratification by type of school. }
}

Output dataset is expressed with ICCAT's coding system.
}
\examples{

# Retrieve ICCAT georeferenced catch data from 2017, with information on fishing mode
iccat_catch<-iccat_catch_level0(2017,include_type_of_school=TRUE)
head(iccat_catch)

}
\seealso{
Other process data: \code{\link{ccsbt_datasets_level0}},
  \code{\link{ccsbt_effort_level0}},
  \code{\link{convert_units}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{iattc_datasets_level0}},
  \code{\link{iccat_effort_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{iotc_effort_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}},
  \code{\link{wcpfc_catch_level0}},
  \code{\link{wcpfc_effort_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
