% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_iccat_functions.R
\name{iccat_functions}
\alias{iccat_functions}
\alias{FUN_catches_ICCAT_CE}
\title{Functions to harmonize the structure of the ICCAT datasets}
\usage{
FUN_catches_ICCAT_CE(RFMO_CE, RFMO_CE_species_colnames)
}
\description{
Set of functions to harmonize the structure of the ICCAT datasets
}
\seealso{
Other harmonize data structure: \code{\link{harmonize_time_structure}},
  \code{\link{iattc_functions}},
  \code{\link{iotc_functions}},
  \code{\link{wcpfc_functions}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
