% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_catch_level0.R
\name{iotc_catch_level0}
\alias{iotc_catch_level0}
\title{Extract source georeferenced catch datasets of IOTC from Sardara World database}
\usage{
iotc_catch_level0(year_tunaatlas)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}
}
\description{
This function extracts the source georeferenced catch datasets stored in the Sardara World database coming from the Indian Ocean Tuna Commission (IOTC). The output dataset provides the catch of tuna, tuna-like and by-catch in the area managed under the jurisdiction of the IOTC. Catches are stratified by month, species, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (usualy 1° or 5° square) and unit of catch (weight or number). Data are expressed using IOTC's coding system.
}
\details{
More information (metadata) regarding the dataset ouput of this function can be found at \href{URLOF DATASET IN CATALOGUE TO ADD}{this URL} on the \href{http://geonetwork2-tunaatlas.d4science.org/geonetwork/srv/fre/catalog.search#/search}{Tuna Atlas catalogue}
}
\examples{

# Retrieve IOTC georeferenced catch data from 2017
iotc_catch<-iotc_catch_level0(2017)
head(iotc_catch)

}
\seealso{
Other extract data: \code{\link{extract_dataset}},
  \code{\link{iotc_effort_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
