% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_effort_level0.R
\name{iotc_effort_level0}
\alias{iotc_effort_level0}
\title{Extract source georeferenced efforts datasets of IOTC from Sardara World database}
\usage{
iotc_effort_level0(year_tunaatlas)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}
}
\description{
This function extracts the source georeferenced effort datasets stored in the Sardara World database coming from the Indian Ocean Tuna Commission (IOTC). The output dataset provides efforts of tuna, tuna-like and shark species in the Indian ocean. Efforts are stratified by month, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (usualy 1° or 5° square) and unit of effort. Data are expressed using IOTC's coding system.
}
\details{

}
\examples{

# Retrieve IOTC georeferenced efforts data from 2017
iotc_effort<-iotc_effort_level0(2017)
head(iotc_effort)

}
\seealso{
Other process data: \code{\link{ccsbt_catch_level0}},
  \code{\link{convert_units}},
  \code{\link{iattc_catch_level0}},
  \code{\link{iccat_catch_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}},
  \code{\link{wcpfc_catch_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
