% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_iotc_functions.R
\name{iotc_functions}
\alias{iotc_functions}
\alias{FUN_catches_IOTC_CE}
\title{Functions to harmonize the structure of the IOTC datasets}
\usage{
FUN_catches_IOTC_CE(Path_to_CE_dataset, last_column_not_catch_value,
  CE_dataset_nature)
}
\description{
Set of functions to harmonize the structure of the IOTC datasets
}
\seealso{
Other harmonize data structure: \code{\link{harmonize_time_structure}},
  \code{\link{iattc_functions}},
  \code{\link{iccat_functions}},
  \code{\link{wcpfc_functions}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
