% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\alias{list_codelists}
\alias{list_raw_datasets}
\alias{list_codelists_mapping}
\title{List metadata of the datasets available in Sardara database}
\usage{
list_datasets(source_authority=NULL,db="sardara_world")
list_raw_datasets(source_authority=NULL,variable=NULL,spatial_resolution=NULL,level_of_correction=NULL,db="sardara_world")
list_codelists<-function(source_authority=NULL,dimension=NULL,db="sardara_world")
list_mappings<-function(source_authority=NULL,dimension=NULL,db="sardara_world")
}
\arguments{
\item{source_authority}{NULL or vector of strings. If not NULL, filter available datasets by the source authority in charge of producing the source statistics collated and harmonized. E.g. c("IOTC","ICCAT") will provide the metadata only for the data produced by IOTC and ICCAT.}

\item{dimension}{NULL or vector of strings. For codelists and mappings only. If not NULL, filter available code lists / mappings by dimensions. E.g. c("gear","species") will provide the metadata only for code lists/mappings between code lists related to fishing gears and species.}

\item{db}{the name of the database to connect to. Defaut connects to Sardara database hosted on the BlueBridge project servers.}

\item{variable}{NULL or vector of strings. For datasets only. If not NULL, filter available code lists / mappings by variable. Three variables are available in Sardara: catch, effort, catch_at_size}

\item{spatial_resolution}{NULL or real. For datasets only. If not NULL, filter available datasets that are defined on that spatial resolution (in degrees).}

\item{level_of_correction}{NULL or integer. For datasets only. If not NULL, filter available datasets that are have that level of correction provided.}
}
\value{
a data.frame of metadata for the data available in Sardara database. For the meaning of the columns of the output data.frame, see \href{https://docs.google.com/spreadsheets/d/1HAgGQzd7GgPXLYgrHECZyqGy4Vsf4IOlppSbmhWEuaE/edit#gid=0}{data dictionary of the metadata table of Sardara database}. ## to fill in this description when it is finalized
}
\description{
This function outputs the metadata of the data available in the Sardara database.
}
\details{
Three types of data are available in Sardara database:
\itemize{
 \item{"raw_dataset": }{Datasets. Three types of datasets are available:
  \itemize{
  \item{"catch": }{Quantity of of fish in number or biomass harvested in a given stratum}
  \item{"effort": }{Quantity of effort (expressed in a given unit such as number of sets, number of fishing hours, number of hooks, etc.) exerted in a given stratum}
  \item{"catch-at-size": }{Quantity of fish at a given size (or size class) harvested in a given stratum}
  }
  }
 \item{"codelist": }{Reference tables with at least codes, and usualy labels associated to these codes.}
 \item{"mapping": }{Tables that establishes mappings / correspondences between codes from various reference tables.}
}
. 
For the meaning of the columns of the raw dataset, see \href{https://docs.google.com/spreadsheets/d/1BUppXu-Z_YX8cJaNISfk9KrKrKJ6y0Dd2XCfIvjBRQc/edit#gid=747135938}{data dictionary of raw datasets}.  ## to fill in this description when it is finalized

\code{list_datasets} lists the metadata of all the types of datasets (raw_dataset, codelists, mappings)
\code{list_raw_datasets} lists only the metadata of raw_datasets
\code{list_codelists} lists only the metadata of the code lists
\code{list_mappings} lists only the metadata of the mappings between code lists
}
\examples{

# List the available source IOTC datasets:
metadata_iotc_datasets<-list_datasets(c("IOTC"))

# List the available code lists for WCPFC and CCSBT
metadata_iccat_code_lists<-list_codelists(c("WCPFC","CCSBT"))

# List the available raw datasets of catch and of effort that are defined on 5° grid resolution
metadata_raw_dataset_catch_5deg<-list_raw_datasets(variable=c("catch","effort"),spatial_resolution=5)


}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
