% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_variable_available_dimensions.R
\name{list_variable_available_dimensions}
\alias{list_variable_available_dimensions}
\title{List dimensions available for a given variable}
\usage{
list_variable_available_dimensions(con,variable)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database)}

\item{variable}{string. The name of a variable (e.g. "catch", or "effort", or "catch_at_size")}
}
\description{
This function returns all the dimensions possibly available for a given variable stored on Sardara DB. NB: the term "variable" is sometimes called "fact".
}
\examples{

#list dimensions available for variable "catch"
list_variable_available_dimensions(db_connection_sardara_world(),"catch")

}
\seealso{
Other discover data: \code{\link{get_codelist_of_dimension}},
  \code{\link{get_dsd_of_dataset}},
  \code{\link{list_dataset_available_dimensions}},
  \code{\link{list_metadata_datasets}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
