% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets_in_db.R
\name{load_datasets_in_db}
\alias{load_datasets_in_db}
\alias{load_raw_dataset_in_db}
\alias{load_df_input_in_db}
\alias{load_mapping_in_db}
\title{Load a dataset in Sardara database}
\usage{
load_raw_dataset_in_db(con,df_to_load,df_metadata,df_codelists_input,variable_name)
load_codelist_in_db(con,df_to_load,df_metadata,dimension_name)
load_mapping_in_db(con,df_to_load,df_metadata)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database) with WRITE permission}

\item{df_to_load}{data.frame to load. See section "Details" for the structure of the dataset}

\item{df_metadata}{data.frame of metadata of to the df_to_load. See section "Details" for the structure of the dataset}

\item{df_codelists_input}{data.frame of the code lists used in case of loading of a raw_dataset. See section "Details" for the structure of the dataset}

\item{variable_name}{string. Name of the variable that is loaded (e.g. "catch", "effort", "catch_at_size")}

\item{dimension_name}{string. Name of the dimension for a code list}
}
\description{
These functions enable to load a dataset in Sardara database. The datasets can be of type "raw_dataset", "df_input" or "mapping"
}
\details{
The data.frames must comply with defined structures. 

For raw datasets (df_to_load used in function load_raw_dataset_in_db): A REDIGER

For code lists (df_to_load used in function load_codelist_in_db): df_to_load should have at minimum 1 column 'code', containing the unique and not null codes for the code list. If there is a column of label, the latter should be named 'labed'. Any additional column is accepted.

For mappings (df_to_load used in function load_mapping_in_db):: The code lists used in the mapping must be available in the DB. df_to_load should have the following structure:
\itemize{
\item{1 column named 'src_code':} {code of the code list source (code to map)}
\item{1 column named 'trg_code':} {code of the code list target}
\item{1 column named 'src_codingsystem':} {name (permanent identifier) of the coding system in Sardara from where the code source is extracted. Only 1 value authorized. The code list must be in the database.}
\item{1 column named 'trg_codingsystem':} {name (permanent identifier) of the coding system in Sardara from where the code target is extracted. The code list must be in the database.}
}

For df_codelists_input: a template is available under inst/extdata/template_df_codelist

For df_metadata: a template of the metadata data.frame to use is available under inst/extdata/template_metadata.csv
}
\examples{


}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
