% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{myfunction}
\alias{myfunction}
\title{myfunction}
\usage{
myfunction(a, b, c)
}
\arguments{
\item{a}{the variable a}

\item{b}{the variable b}

\item{c}{the variable c}
}
\value{
an object of class "numeric"
}
\description{
\code{myfunction} is the function to perform ....
}
\examples{
  out <- myfunction(a = 1, b = 2, c = 3)  

}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
