% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_map.R
\name{pie_map}
\alias{pie_map}
\title{Create a pie map}
\usage{
pie_map(con, df_input, dimension_group_by, df_spatial_code_list_name,
  area_filter_wkt = NULL, number_of_classes)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database)}

\item{df_input}{data.frame to map}

\item{dimension_group_by}{string. Name of the dimension that will be the classes in the pies}

\item{df_spatial_code_list_name}{string. Name of the spatial coding system used in df_input}

\item{number_of_classes}{integer. Number of classes to visualize in the pies.}
}
\description{
This function ouputs a pie map from the dataset provided as input
}
\details{

}
\examples{

# Connect to Sardara DB
con <- db_connection_sardara_world()

# Extract IOTC (Indian Ocean) georeferenced catch time series of catches from Sardara DB, in 5° resolution
ind_catch_tunaatlasird_level2<-extract_dataset(con,list_metadata_datasets(con,dataset_name="indian_ocean_catch_5deg_1m_1952_11_01_2016_01_01_tunaatlasIRD_level2"))
head(ind_catch_tunaatlasird_level2)

# filter the data to keep only catches on log schools in 2014:
ind_catch_tunaatlasird_level2 <- ind_catch_tunaatlasird_level2 \%>\% filter (year==2014) \%>\% filter (schooltype=="LS")

# Map the catches made on log schools in 2014 by species:
PieMap(con,
      df_input=ind_catch_tunaatlasird_level2,
      dimension_group_by="species",
      df_spatial_code_list_name="areas_tuna_rfmos_task2",
      number_of_classes=4 
     )
     
dbDisconnect(con)
}
\seealso{
Other visualize data: \code{\link{functions_visualize_data}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
