% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_geo_timeseries.R
\name{rasterize_geo_timeseries}
\alias{rasterize_geo_timeseries}
\title{Aggregate a geospatial time series on polygons}
\usage{

}
\arguments{
\item{df_input}{}

\item{intersection_layer}{sf polygon}

\item{calendar}{}

\item{buffer}{real only in case df_input is a trajectory}
}
\value{

}
\details{
df_input must have columns: date, lat, lon
if df_input is a trajectory, must have 2 additional columns: id_object and id_trajectory

calendar must have columns: time_start and time_end

intersection_layer: type sp SpatialPolygonsDataFrame

The input data that do not intersect any polygon of the intersection_layer are marked as 'no_geo_intersection' in the column 'geographic_identifier' of the output dataset
}
\examples{

# Create a continous calendar of 1 month time step
calendar<-create_calendar("2000-01-01","2020-01-01",1,"month")
# Create a ĝrid of 1° resolution
grid <- create_grid(latmin=-90,latmax=90,lonmin=-180,lonmax=80,resolution=1,centred=T)

# Rasterize


}
\seealso{
Other process data: \code{\link{convert_units}},
  \code{\link{create_calendar}}, \code{\link{create_grid}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}}
}
\author{
Chloé Dalleau, \email{dalleau.chloe@hotmail.fr}; modified by Paul Taconet, \email{paul.taconet@ird.fr}
}
