% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtunaatlas.R
\docType{package}
\name{rtunaatlas}
\alias{rtunaatlas}
\alias{rtunaatlas-package}
\title{IRD-Sardara Tuna Atlas R facilities}
\description{
IRD-Sardara Tuna Atlas R Facilities
}
\details{
\tabular{ll}{
   Package: \tab rtunaatlas\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2017-06-23\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Paul Taconet \email{paul.taconet@ird.fr}
}

