% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_curation_intersect_areas.R
\name{spatial_curation_intersect_areas}
\alias{spatial_curation_intersect_areas}
\title{Perform spatial intersection with a geospatial layer}
\usage{
spatial_curation_intersect_areas(
con,
df_input,
df_spatial_code_list_name,
intersection_spatial_code_list_name)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database)}

\item{df_input}{data.frame of fact. The dataframe must have at least a column 'geographic_identifier'. The code list used for the column 'geographic_identifier' must be one of the table of the schema 'area' of the database.}

\item{df_spatial_code_list_name}{string . Dataset name of the code list used for the column 'geographic_identifier' of \code{df_input} (type polygon)}

\item{intersection_spatial_code_list_name}{string .  Name of the intersection layer to use (type polygon). Corresponds to the name of the table in the database.}
}
\value{
a list with 2 objects: 
\itemize{
 \item{"df": }{\code{df_input} with additional columns:
 \itemize{
 \item{"geographic_identifier_intersection_layer": }{ Code of the feature from the intersection layer (\code{intersection_spatial_code_list_name}) that intersects the geographical identifier of \code{df_input}}
 \item{"proportion_source_area_intersection": }{ Proportion of the area of the \code{df_input} feature being intersected by the \code{intersection layer} feature. }
 }}
 \item{"df_input_areas_intersect_intersection_layer": }{a data.frame summarizing the results of the spatial intersection. The columns are: }
 \itemize{
 \item{"geographic_identifier_source_layer": }{ Code of the feature (geographic identifier) from \code{df_input}}
 \item{"geographic_identifier_intersection_layer": }{ Code of the feature from (\code{intersection_spatial_code_list_name}) that intersects the feature of \code{df_input}}
 \item{"codelist_source_layer": }{ Name of the code list used for the column 'geographic_identifier' of \code{df_input}}
 \item{"codelist_intersection_layer": }{ Name of the intersection layer code list (\code{intersection_spatial_code_list_name}) }
 \item{"proportion_source_area_intersection": }{ Proportion of the area of the \code{input layer} feature being intersected by the \code{intersection layer} feature. }
 }

}
}
\description{
This function performs the spatial intersection between a georeferenced dataset (\code{input layer}) and a georeferenced code list stored in Sardara database (\code{intersection layer}). It provides the features of the \code{intersection layer} that intersect the features of the \code{input layer}, along with the proportion of the area of the \code{input layer} feature being intersected by the \code{intersection layer} feature.
}
\details{
User has to be aware that the intersection might return several rows for 1 given row of \code{df_input}. This happens when one spatial feature of the \code{input layer} intersects several features of the \code{intersection layer}. Hence, the values of the ouput dataset might not be summed directly.
Example:

1 row of \code{df_input}:
\tabular{rrrrrrr}{
flag	\tab time_start	\tab time_end	  \tab  geographic_identifier 	\tab  gear	 \tab  species	 \tab  value  \cr
AUS	\tab 1992-02-01	\tab 1992-03-01	\tab 235140	\tab LL	\tab YFT	\tab 0.05 \cr
}


When executing the function \code{function spatial_curation_intersect_areas} taking as intersection layer "fao_fishing_areas", the output dataset will return 2 rows:


\tabular{rrrrrrrrr}{
flag	 \tab  time_start 	\tab  time_end	  \tab  geographic_identifier	 \tab  gear	 \tab  species	\tab  value  \tab  geographic_identifier_intersection_layer  \tab  proportion_intersection  \cr 
AUS	\tab 1992-02-01	\tab 1992-03-01	\tab 235140	\tab LL	\tab YFT	\tab 3.2 \tab F51 \tab 0.5 \cr 
AUS	\tab 1992-02-01	\tab 1992-03-01	\tab 235140	\tab LL	\tab YFT	\tab 3.2 \tab F57 \tab 0.5 \cr 
}

This table means that the geograpical identifier "235140" intersects two areas of the intersection layer: "F51" with 50\% (proportion_intersection=0.5) of the area of 235140 being intersected by F51, and "F57" with 50\% (proportion_intersection=0.5) of the area of 235140 being intersected by F57. However, the value of catch is 3.2 in both cases - i.e. it has not been multiplied by the proportion of area intersected.
}
\examples{

# Connect to Tuna atlas database
con<-db_connection_tunaatlas_world()

# Extract a dataset
dataset_metadata<-list_metadata_datasets(con,identifier="west_pacific_ocean_catch_1970_01_01_2016_01_01_5deg_1m_ps_tunaatlasWCPFC_2017_level0")
df<-extract_dataset(con,dataset_metadata)

# Retrieve the spatial coding system used in the dataset
df_spatial_codingsystem <- get_codelist_of_dimension(con,dataset_metadata,"area")

# Get the list of geospatial layers available in the database
spatial_code_lists_available<-list_metadata_codelists(con,dimension="area")

# Intersect df with continents (to check which data are located on land areas)
df_intersect_continents<-spatial_curation_intersect_areas(con,df_input=df,df_spatial_code_list_name=df_spatial_codingsystem$dataset_name,intersection_spatial_code_list_name="gshhs_world_coastlines")

head(df_intersect_continents$df)
head(df_intersect_continents$df_input_areas_intersect_intersection_layer)

# Intersect df with EEZ (to know what percentage of df is located on EEZs)
df_intersect_eez<-spatial_curation_intersect_areas(con,df_input=df,df_spatial_code_list_name=df_spatial_codingsystem$dataset_name,intersection_spatial_code_list_name="vliz_world_eez_v8_2014")

head(df_intersect_eez$df)
head(df_intersect_eez$df_input_areas_intersect_intersection_layer)

}
\seealso{
Other process data: \code{\link{convert_units}},
  \code{\link{create_calendar}}, \code{\link{create_grid}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{rasterize_geo_timeseries}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
