% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_curation_reallocate_data.R
\name{spatial_curation_reallocate_data}
\alias{spatial_curation_reallocate_data}
\alias{spatial_curation_function_reallocate_data}
\title{Reallocate data using a spatial criterium}
\usage{
spatial_curation_function_reallocate_data(df_input, dimension_reallocation,
  vector_to_reallocate, reallocation_dimensions)
}
\arguments{
\item{df_input}{data.frame of fact}

\item{dimension_reallocation}{dimension to reallocate}

\item{vector_to_reallocate}{vector of data to reallocate}

\item{reallocation_dimensions}{vector of dimensions to use for the reallocation}
}
\value{
a list with 2 objects: 
 \itemize{
 \item{"df": }{\code{df_input} where data have been reallocated}
 \item{"stats": }{A data.frame with some information regarding the reallocation process:}
   \itemize{
    \item{"value": }{sum in the input dataset}
    \item{"value_reallocate": }{sum reallocated}
   }
 }
}
\description{
This function reallocates data
}
\details{
Example of use of this function: reallocate the catch of the areas that are located in land over the areas that are located at sea with the same dimensions.
}
\examples{


# Extract a dataset from Sardara World database
con=db_connection_sardara_world()
dataset_metadata<-list_metadata_datasets(con,dataset_name="atlantic_ocean_catch_1950_01_01_2016_01_01_tunaatlasICCAT_2017_level0__noSchool")
df<-extract_dataset(con,dataset_metadata)

# Retrieve the spatial coding system used in the dataset
df_spatial_codingsystem <- get_codelist_of_dimension(con,dataset_metadata,"area")

# Get data that are fully located inland:
df_intersect_continents<-spatial_curation_intersect_areas(con,df_input=df,df_spatial_code_list_name=df_spatial_codingsystem$dataset_name,intersection_spatial_code_list_name="gshhs_world_coastlines")
areas_inland<-df_intersect_continents$df_input_areas_intersect_intersection_layer$geographic_identifier_source_layer[which(df_intersect_continents$df_input_areas_intersect_intersection_layer$proportion_source_area_intersection==1)]
# In the above, df_intersect_continents$df_input_areas_intersect_intersection_layer$proportion_source_area_intersection==1 means that we want to extract the codes of the areas that are located fully inland (100\% of the surface is located inland). 

# Reallocate data that are located inland. Reallocate on strata with same month, gear, flag, species, unit :
dataset_reallocate_data_inland<-spatial_curation_function_reallocate_data(
df_input=df,
dimension_reallocation="geographic_identifier",
vector_to_reallocate=areas_inland,
reallocation_dimensions=c("month","gear","flag","species","unit"))

# New dataset with data that were located inland reallocated to areas at sea with same dimensions
head(dataset_reallocate_data_inland$df)

# To get some information on the reallocation process:
dataset_reallocate_data_inland$stats

}
\seealso{
Other process data: \code{\link{ccsbt_catch_level0}},
  \code{\link{convert_units}},
  \code{\link{iattc_catch_level0}},
  \code{\link{iccat_catch_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{iotc_effort_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_upgrade_resolution}},
  \code{\link{wcpfc_catch_level0}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
