% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_curation_upgrade_resolution.R
\name{spatial_curation_upgrade_resolution}
\alias{spatial_curation_upgrade_resolution}
\title{Aggregate gridded data}
\usage{
spatial_curation_upgrade_resolution(con, df_input, resolution)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database)}

\item{df_input}{data.frame of fact. The data frame must contain at least one column "geographical_identifier" with CWP grid codification or IOTC irregular areas from catch-and-effort datasets.}

\item{resolution}{integer. Resolution to reach (in degrees). Currently, works only with 5.}
}
\value{
a list with 2 objects: 
 \itemize{
 \item{"df": }{\code{df_input} where data have been aggregated}
 \item{"stats": }{A data.frame with some information regarding the process}
 }
}
\description{
This function aggregates the data of a dataset on a grid with resolution equal to \code{resolution}. Data with resolutions inferior to \code{resolution} will be aggregated on the corresponding \code{resolution} quadrant. Data with resolutions superior to \code{resolution} will not be disaggregated. To disaggregate data with resolutions superior to \code{resolution}, use the function \code{spatial_curation_downgrade_resolution}
}
\examples{


# Extract a df_input from Sardara World database
con=db_connection_sardara_world()
dataset_metadata<-list_metadata_datasets(con,dataset_name="atlantic_ocean_catch_1950_01_01_2016_01_01_tunaatlasICCAT_2017_level0__noSchool")
df<-extract_dataset(con,dataset_metadata)

# Aggregate data with resolutions inferior to 5° on the corresponding 5° resolution grid:
df_aggregated_5_deg<-spatial_curation_upgrade_resolution(con,df_input=df,resolution=5)

head(df_aggregated_5_deg$df)
# In this dataset, data that were defined on a grid inferior to 5° were aggregated on the corresponding 5° resolution, other data remained at their source resolution
# if you want the data superior to 5° to be disaggregated on the 5° grid, use : 
# spatial_curation_downgrade_resolution(con,df_input=df_aggregated_5_deg,resolution=5)

}
\seealso{
Other process data: \code{\link{convert_units}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
