% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcpfc_catch_level0.R
\name{wcpfc_catch_level0}
\alias{wcpfc_catch_level0}
\title{Extract source georeferenced catch datasets of WCPFC from Sardara World database}
\usage{
wcpfc_catch_level0(year_tunaatlas)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}
}
\description{
This function extracts the source georeferenced catch datasets stored in the Sardara World database coming from the Western and Central Pacific Ocean Fisheries Commission (WCPFC). The output dataset provides the catch of tuna and tuna-like species in the area managed under the jurisdiction of the WCPFC. Catches are stratified by month, species, gear, fishing mode (i.e. type of school used), area (5° square) and unit of catch (weight or number). Data are expressed using WCPFC's coding system.
}
\details{
The output dataset lists catch of tuna and tuna-like species in the Western Pacific ocean. Catches are stratified by month, species, gear, fishing mode (i.e. type of school used), area (5° square) and unit of catch (weight or number). The dataset is computed using public domain datasets released by the Western and Central Pacific Fisheries Commission (WCPFC).
This function merges the primary catch-and-effort datasets released by WCPFC.
Data are expressed using WCPFC's coding system.
}
\examples{

# Retrieve WCPFC georeferenced catch data from 2017
wcpfc_catch<-wcpfc_catch_level0(2017)
head(wcpfc_catch)

}
\seealso{
Other extract data: \code{\link{ccsbt_catch_level0}},
  \code{\link{extract_dataset}},
  \code{\link{iccat_catch_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{iotc_effort_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
