% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcpfc_effort_level0.R
\name{wcpfc_effort_level0}
\alias{wcpfc_effort_level0}
\title{Extract source georeferenced efforts datasets of WCPFC from the Tuna atlas database}
\usage{
wcpfc_effort_level0(year_tunaatlas)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}
}
\description{
This function extracts the source georeferenced effort datasets stored in the Tuna atlas database coming from the Western and Central Pacific Ocean Fisheries Commission (WCPFC). The output dataset provides efforts of tuna, tuna-like and shark species in the Western pacific ocean. Efforts are stratified by year, month, fishing mode, unit of effort, area (5° squares). Data are expressed using WCPFC's coding system.
}
\details{

}
\examples{

# Retrieve wcpfc georeferenced efforts data from 2017
wcpfc_effort<-wcpfc_effort_level0(2017)
head(wcpfc_effort)

}
\seealso{
Other process data: \code{\link{ccsbt_datasets_level0}},
  \code{\link{ccsbt_effort_level0}},
  \code{\link{convert_units}},
  \code{\link{get_rfmos_datasets_level0}},
  \code{\link{iattc_datasets_level0}},
  \code{\link{iccat_catch_level0}},
  \code{\link{iccat_effort_level0}},
  \code{\link{iotc_catch_level0}},
  \code{\link{iotc_effort_level0}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}},
  \code{\link{wcpfc_catch_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
