% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_wcpfc_functions.R
\name{wcpfc_functions}
\alias{wcpfc_functions}
\alias{FUN_catches_WCPFC_CE_allButPurseSeine}
\title{Functions to harmonize the structure of the WCPFC datasets}
\usage{
FUN_catches_WCPFC_CE_allButPurseSeine(DF_Path)
}
\description{
Set of functions to harmonize the structure of the WCPFC datasets
}
\seealso{
Other harmonize data structure: \code{\link{harmonize_time_structure}},
  \code{\link{iattc_functions}},
  \code{\link{iccat_functions}},
  \code{\link{iotc_functions}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
