% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datasets.R
\name{extract_dataset}
\alias{extract_dataset}
\alias{extract_datasets}
\title{Extract a dataset available in Sardara database}
\usage{
extract_dataset(con,dataset_name)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database)}

\item{dataset_name}{string. The name of a dataset in the database (column "dataset_name" of the table metadata.metadata)}
}
\value{
a data.frame of the data available in the database and set as \code{dataset_name}
}
\description{
This function outputs a dataset available in the Sardara database.
}
\details{
The meaning of the columns of a raw dataset is provided in the \href{https://docs.google.com/spreadsheets/d/1BUppXu-Z_YX8cJaNISfk9KrKrKJ6y0Dd2XCfIvjBRQc/edit#gid=747135938}{data dictionary of raw datasets}.  ## to fill in this description when it is finalized
}
\examples{

# Extract a raw dataset
raw_dataset<-extract_dataset(con,"global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1")

# Extract a code list
code_list<-extract_dataset(con,"species_asfis")

# Extract a mapping between code lists
code_list_mapping<-extract_dataset(con,"codelist_mapping_species_iotc_speciesgroup_tunaatlas")


}
\seealso{
Other extract data: \code{\link{iotc_catch_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
