% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rfmos_datasets_level0.R
\name{get_rfmos_datasets_level0}
\alias{get_rfmos_datasets_level0}
\title{Extract primary gridded tuna RFMOs datasets from the Tuna atlas database}
\usage{
get_rfmos_datasets_level0(rfmo,variable,year_tunaatlas,iattc_raise_flags_to_schooltype,iattc_dimension_to_use_if_no_raising_flags_to_schooltype,iccat_include_type_of_school)
}
\arguments{
\item{rfmo}{string. Acronym of the RFMO. Accepted values: "IOTC", "ICCAT", "IATTC", "WCPFC", "CCSBT". See additional parameters to set if \code{rfmo} is set to "IATTC" or "ICCAT"}

\item{variable}{string. Variable to extract. Accepted values: "catch", "effort"}

\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}

\item{iattc_raise_flags_to_schooltype}{boolean. Use only if \code{rfmo}=="IATTC". Raise dataset with flag stratification to dataset with schooltype stratification? See section Details for more information.}

\item{iattc_dimension_to_use_if_no_raising_flags_to_schooltype}{NULL or string. Use only if \code{rfmo}=="IATTC". Not nullable if \code{raise_flags_to_schooltype} is set to FALSE. if not NULL, either set to "flag" or "schooltype"}

\item{iccat_include_type_of_school}{boolean. Use only if \code{rfmo}=="ICCAT". Set TRUE if you want the output dataset with school type stratification. FALSE will provide the stratification without the type of school. See details for more information.}
}
\description{
This function extracts the primary gridded time series coming from the tuna Regional fisheries management organizations and stored in the Tuna atlas database. Data include geo-spatial gridded catch and efforts.
}
\details{
The output dataset lists catch or effort of tuna, tuna-like and shark species in the area of competence of the RFMO specified. Catches or efforts are usually stratified by year, month, species (for catches only), fishing gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (1° or 5° square) and unit. Some of these dimensions can be missing depending on the confidentialy policies of each RFMO. The output dataset is computed using public domain datasets released by the RFMOs.

Details on the use of the parameter \code{iattc_raise_flags_to_schooltype}: For confidentiality policies, information on flag and school type for the geo-referenced catches is available in separate files for IATTC Purse seine datasets.
\itemize{
\item{ If the parameter \code{iattc_raise_flags_to_schooltype} is set to TRUE, for each stratum, the catch/effort from the flag-detailed dataset will be raised to the catch/effort from the school type-detailed dataset to get an estimation of the catches by flag and school type in each stratum.}
\item{ If the parameter \code{iattc_raise_flags_to_schooltype} is set to FALSE, one single dataset will be used and in this case, the parameter \code{dimension_to_use_if_no_raising_flags_to_schooltype} must be filled in: }
 \itemize{
\item{ If the parameter \code{iattc_dimension_to_use_if_no_raising_flags_to_schooltype} is set to "flag", the data with flag information will be used.}
\item{ If the parameter \code{iattc_dimension_to_use_if_no_raising_flags_to_schooltype} is set to "schooltype", the data with schooltype information will be used. }
}
}

#' Details on the use of the parameter \code{iccat_include_type_of_school}: ICCAT delivers two catch-and-efforts datasets for purse seiners: one that gives the detail of the type of school (Fad|Free school) for purse seine fisheries and that starts in 1994 (called Task II catch|effort by operation mode Fad|Free school) and one that does not give the information of the type of school and that covers all the time period (from 1950) (called Task II catch|effort). These data are redundant (i.e. the data from the dataset Task II catch|effort by operation mode are also available in the dataset Task II catch|effort) but in the latter, the information on the type of school is not available.
\itemize{
\item{ If the parameter \code{iccat_include_type_of_school} is set to TRUE, both datasets will be combined to produce a dataset that covers the whole time period, with fishing mode information (Fad | free school).}
\item{ If the parameter \code{iccat_include_type_of_school} is set to FALSE, only the dataset without the type of school information will be used. The output dataset will hence not have the stratification by type of school. }
}

The output dataset is expressed with the RFMO's coding system.
}
\examples{

# Retrieve IATTC georeferenced catch data from 2017, with dataset with flag dimension raised to dataset with schoolytpe dimension
iattc_catch<-get_rfmos_datasets_level0("iattc","catch",2017,raise_flags_to_schooltype=TRUE)
head(iattc_catch)

}
\seealso{
Other process data: \code{\link{convert_units}},
  \code{\link{map_codelist}},
  \code{\link{raise_datasets_by_dimension}},
  \code{\link{raise_get_rf}},
  \code{\link{raise_incomplete_dataset_to_total_dataset}},
  \code{\link{spatial_curation_downgrade_resolution}},
  \code{\link{spatial_curation_intersect_areas}},
  \code{\link{spatial_curation_reallocate_data}},
  \code{\link{spatial_curation_upgrade_resolution}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
