% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSQLSardaraQueries.R
\name{getSQLSardaraQueries}
\alias{getSQLSardaraQueries}
\title{Get a list of queries to execute to plug services to the datasets available in Sardara}
\usage{
getSQLSardaraQueries(con,dataset_name)
}
\arguments{
\item{con}{a wrapper of rpostgresql connection (connection to a database)}

\item{dataset_name}{string. The name of a dataset in the database (column "dataset_name" of the table metadata.metadata)}
}
\value{
a list with SQL queries and informations on the dataset. Depending on the type of data (raw_dataset, codelist, mapping), the output elements of the list might vary. The elements of the list are:
\itemize{
 \item{\code{query_CSV} :}{ Query to retrieve a data.frame }
 \item{\code{query_NetCDF} :}{ }
 \item{\code{query_wfs_wms} :}{ }
 \item{\code{query_wfs_wms_aggregated_layer} :}{ }
 \item{\code{query_dynamic_list_keywords_species} :}{ }
 \item{\code{query_dynamic_list_keywords_fishing_gears} :}{ }
 \item{\code{query_dynamic_list_keywords_Fleets} :}{ }
 \item{\code{query_dynamic_metadata_count_features} :}{ }
 \item{\code{query_dynamic_metadata_spatial_Extent} :}{ }
 \item{\code{query_dynamic_metadata_temporal_Extent} :}{ }
 \item{\code{query_dynamic_metadata_get_SRID} :}{ }
 \item{\code{lineage} :}{ }
 \item{\code{query_dynamic_list_keywords_institutions} :}{ }
  }
}
\description{
Get a list of queries to execute to plug services to the datasets available in Sardara
}
\examples{

queries<-getSQLSardaraQueries(db_connection_sardara_world(),"global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1")

# retrieve data.frame of global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1

global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1<-dbGetQuery(con,queries$query_CSV)

}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
