% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_catch_level0.R
\name{iotc_catch_level0}
\alias{iotc_catch_level0}
\title{Extract source georeferenced catch datasets of IOTC from Sardara World database}
\usage{
iotc_catch_level0(year_tunaatlas)
}
\arguments{
\item{year_tunaatlas}{numeric. The year of the datasets to extract. Starts in 2017}
}
\description{
This function extracts the source georeferenced catch datasets stored in the Sardara World database coming from the Indian Ocean Tuna Commission (IOTC). The output dataset provides the catch of tuna, tuna-like and by-catch in the area managed under the jurisdiction of the IOTC. Catches are stratified by month, species, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (usualy 1째 or 5째 square) and unit of catch (weight or number).
}
\details{
The output dataset lists catch of tuna, tuna-like and shark species in the Indian ocean. Catches are stratified by month, species, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (1째 or 5째 square) and unit of catch (weight or number). This dataset is computed using public domain datasets released by the Indian Ocean Tuna Commission (IOTC).
This function merges the primary catch-and-effort datasets released by IOTC.
Output dataset is expressed with IOTC's coding system.
}
\examples{

# Retrieve IOTC georeferenced catch data from 2017
iotc_catch<-iotc_catch_level0(2017)
head(iotc_catch)

}
\seealso{
Other tRFMOs datasets extraction: \code{\link{ccsbt_catch_level0}},
  \code{\link{iattc_catch_level0}},
  \code{\link{iccat_catch_level0}},
  \code{\link{wcpfc_catch_level0}}

Other datasets extraction: \code{\link{ccsbt_catch_level0}},
  \code{\link{extract_dataset}},
  \code{\link{iattc_catch_level0}},
  \code{\link{iccat_catch_level0}},
  \code{\link{wcpfc_catch_level0}}
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
