% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\alias{list_raw_datasets}
\alias{list_codelists}
\alias{list_codelists_mapping}
\title{List metadata of the datasets available in Sardara database}
\usage{
list_datasets(source_authority=NULL,db="sardara_world")
list_raw_datasets(source_authority=NULL,variable=NULL,spatial_resolution=NULL,level_of_correction=NULL,db="sardara_world")
list_codelists(source_authority=NULL,dimension=NULL,db="sardara_world")
list_mappings(source_authority=NULL,dimension=NULL,db="sardara_world")
}
\arguments{
\item{source_authority}{NULL or vector of strings. If not NULL, filter available datasets by the source authority in charge of producing the source statistics collated and harmonized. E.g. c("IOTC","ICCAT") will provide the metadata only for the data produced by IOTC and ICCAT.}

\item{db}{the name of the database to connect to. Defaut connects to Sardara database hosted on the BlueBridge project servers.}

\item{dimension}{NULL or vector of strings. For codelists and mappings only. If not NULL, filter available code lists / mappings by dimensions. E.g. c("gear","species") will provide the metadata only for code lists/mappings between code lists related to fishing gears and species.}

\item{variable}{NULL or vector of strings. For datasets only. If not NULL, filter available code lists / mappings by variable. Three variables are available in Sardara: catch, effort, catch_at_size}

\item{spatial_resolution}{NULL or real. For datasets only. If not NULL, filter available datasets that are defined on that spatial resolution (in degrees).}

\item{level_of_correction}{NULL or integer. For datasets only. If not NULL, filter available datasets that are have that level of correction provided.}
}
\value{
a data.frame of metadata for the data available in Sardara database. The meaning of the columns of the output data.frame is provided in the \href{https://docs.google.com/spreadsheets/d/1HAgGQzd7GgPXLYgrHECZyqGy4Vsf4IOlppSbmhWEuaE/edit#gid=0}{data dictionary of the metadata table of Sardara database}. ## to fill in this description when it is finalized
}
\description{
This function outputs the metadata of the data available in the Sardara database.
}
\examples{

# List the available source IOTC datasets:
metadata_iotc_datasets<-list_datasets(c("IOTC"))

# List the available code lists for WCPFC and IATTC
metadata_iccat_code_lists<-list_codelists(c("WCPFC","IATTC"))

# List the available raw datasets of catch and of effort that are defined on 5° grid resolution
metadata_raw_dataset_catch_5deg<-list_raw_datasets(variable=c("catch","effort"),spatial_resolution=5)


}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
